move over binary_length
