+++ daft/daft/__init__.pyi
@@ -1276,7 +1276,6 @@ def list_chunk(expr: PyExpr, size: int) -> PyExpr: ...
 # ---
 # expr.binary namespace
 # ---
-def binary_length(expr: PyExpr) -> PyExpr: ...
 def binary_slice(expr: PyExpr, start: PyExpr, length: PyExpr | None = None) -> PyExpr: ...
 
 class PyCatalog:
