+++ src/daft-functions-binary/src/lib.rs
@@ -3,6 +3,7 @@ mod concat;
 mod decode;
 mod encode;
 mod kernels;
+mod length;
 mod utils;
 
 use daft_dsl::functions::{FunctionModule, FunctionRegistry};
@@ -16,5 +17,6 @@ impl FunctionModule for BinaryFunctions {
         parent.add_fn(decode::BinaryTryDecode);
         parent.add_fn(encode::BinaryEncode);
         parent.add_fn(encode::BinaryTryEncode);
+        parent.add_fn(length::BinaryLength);
     }
 }
