+++ src/daft-functions/src/binary/mod.rs
@@ -1,5 +1,3 @@
-pub mod length;
 pub mod slice;
 
-pub use length::{binary_length, BinaryLength};
 pub use slice::{binary_slice, BinarySlice};
